# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/quizzes": {
      "get": {
        "operationId": "QuizzesController_all",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC. Accepted: SUPER_USER, ADMIN, INSTRUCTOR, COURSE_PROVIDER, LEARNER. Lowercase frontend aliases also work: superuser, admin, instructor, provider, learner.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "SUPER_USER",
                "ADMIN",
                "INSTRUCTOR",
                "COURSE_PROVIDER",
                "LEARNER",
                "superuser",
                "admin",
                "instructor",
                "provider",
                "learner"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "Optional current user id used by the frontend session. Useful for owner checks and demo traceability.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "u6"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "403": {
            "description": "Forbidden: missing or unauthorized role header"
          }
        },
        "tags": [
          "Quizzes"
        ]
      },
      "post": {
        "operationId": "QuizzesController_create",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC. Accepted: SUPER_USER, ADMIN, INSTRUCTOR, COURSE_PROVIDER, LEARNER. Lowercase frontend aliases also work: superuser, admin, instructor, provider, learner.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "SUPER_USER",
                "ADMIN",
                "INSTRUCTOR",
                "COURSE_PROVIDER",
                "LEARNER",
                "superuser",
                "admin",
                "instructor",
                "provider",
                "learner"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "Optional current user id used by the frontend session. Useful for owner checks and demo traceability.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "u6"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuizDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "403": {
            "description": "Forbidden: missing or unauthorized role header"
          }
        },
        "tags": [
          "Quizzes"
        ]
      }
    },
    "/api/quizzes/{id}": {
      "get": {
        "operationId": "QuizzesController_one",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC. Accepted: SUPER_USER, ADMIN, INSTRUCTOR, COURSE_PROVIDER, LEARNER. Lowercase frontend aliases also work: superuser, admin, instructor, provider, learner.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "SUPER_USER",
                "ADMIN",
                "INSTRUCTOR",
                "COURSE_PROVIDER",
                "LEARNER",
                "superuser",
                "admin",
                "instructor",
                "provider",
                "learner"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "Optional current user id used by the frontend session. Useful for owner checks and demo traceability.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "u6"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "403": {
            "description": "Forbidden: missing or unauthorized role header"
          }
        },
        "tags": [
          "Quizzes"
        ]
      },
      "patch": {
        "operationId": "QuizzesController_update",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC. Accepted: SUPER_USER, ADMIN, INSTRUCTOR, COURSE_PROVIDER, LEARNER. Lowercase frontend aliases also work: superuser, admin, instructor, provider, learner.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "SUPER_USER",
                "ADMIN",
                "INSTRUCTOR",
                "COURSE_PROVIDER",
                "LEARNER",
                "superuser",
                "admin",
                "instructor",
                "provider",
                "learner"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "Optional current user id used by the frontend session. Useful for owner checks and demo traceability.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "u6"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateQuizDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          },
          "403": {
            "description": "Forbidden: missing or unauthorized role header"
          }
        },
        "tags": [
          "Quizzes"
        ]
      },
      "delete": {
        "operationId": "QuizzesController_remove",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC. Accepted: SUPER_USER, ADMIN, INSTRUCTOR, COURSE_PROVIDER, LEARNER. Lowercase frontend aliases also work: superuser, admin, instructor, provider, learner.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "SUPER_USER",
                "ADMIN",
                "INSTRUCTOR",
                "COURSE_PROVIDER",
                "LEARNER",
                "superuser",
                "admin",
                "instructor",
                "provider",
                "learner"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "Optional current user id used by the frontend session. Useful for owner checks and demo traceability.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "u6"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "403": {
            "description": "Forbidden: missing or unauthorized role header"
          }
        },
        "tags": [
          "Quizzes"
        ]
      }
    },
    "/api/quizzes/{id}/submit": {
      "post": {
        "operationId": "QuizzesController_submit",
        "summary": "Submit quiz answers",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC. Accepted: SUPER_USER, ADMIN, INSTRUCTOR, COURSE_PROVIDER, LEARNER. Lowercase frontend aliases also work: superuser, admin, instructor, provider, learner.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "SUPER_USER",
                "ADMIN",
                "INSTRUCTOR",
                "COURSE_PROVIDER",
                "LEARNER",
                "superuser",
                "admin",
                "instructor",
                "provider",
                "learner"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "Optional current user id used by the frontend session. Useful for owner checks and demo traceability.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "u6"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitQuizDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "403": {
            "description": "Forbidden: missing or unauthorized role header"
          }
        },
        "tags": [
          "Quizzes"
        ]
      }
    },
...
  },
  "info": {
    "title": "IndAI Review-4 Paper Trading API",
    "description": "NestJS modular backend with in-memory arrays, RBAC through x-role header, DTO validation, Swagger and minute-stock data endpoints.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateQuizDto": {
        "type": "object",
        "properties": {
          "courseId": {
            "type": "string",
            "example": "c1"
          },
          "providerId": {
            "type": "string",
            "example": "u5"
          },
          "title": {
            "type": "string",
            "example": "Basics Quiz",
            "minLength": 3
          },
          "questions": {
            "type": "object",
            "description": "Frontend may send a question count number; API may also accept question objects.",
            "example": [
              {
                "q": "Paper trading?",
                "options": [
                  "Virtual",
                  "Real"
                ],
                "ans": 0
              }
            ]
          },
          "questionCount": {
            "type": "number",
            "example": 10,
            "minimum": 1,
            "maximum": 50
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive"
            ]
          }
        },
        "required": [
          "title"
        ]
      },
      "UpdateQuizDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 3
          },
          "courseId": {
            "type": "string"
          },
          "questions": {
            "type": "object"
          },
          "questionCount": {
            "type": "number",
            "minimum": 1,
            "maximum": 50
          },
          "timeLimit": {
            "type": "number",
            "minimum": 5,
            "maximum": 120
          },
          "passingScore": {
            "type": "number",
            "minimum": 1,
            "maximum": 100
          },
          "skillPointsOnPass": {
            "type": "number",
            "minimum": 0,
            "maximum": 100
          },
          "instructions": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive"
            ]
          }
        }
      },
      "SubmitQuizDto": {
        "type": "object",
        "properties": {
          "learnerId": {
            "type": "string",
            "example": "u6"
          },
          "answers": {
            "example": [
              0,
              1,
              2
            ],
            "type": "array",
            "items": {
              "type": "number"
            }
          }
        },
        "required": [
          "learnerId",
          "answers"
        ]
      },
...
    }
  }
}