# back-end/src/modules/quizzes/quizzes.controller.ts
import { BadRequestException, Body, Controller, Delete, Get, NotFoundException, Param, Patch, Post } from '@nestjs/common';
import { ApiBody, ApiOperation, ApiTags } from '@nestjs/swagger';
import { DataStore } from '../../store/data.store';
import { Roles } from '../../common/roles.decorator';
import { ApiRoleHeader } from '../../common/swagger-role-header';
import { CreateQuizDto, SubmitQuizDto, UpdateQuizDto } from './dto';
@ApiTags('Quizzes') @Controller('quizzes') @ApiRoleHeader()
export class QuizzesController { constructor(private readonly db:DataStore){}
 @Get() @Roles('superuser','admin','instructor','provider','learner') all(){return {success:true,data:this.db.quizzes};}
 @Get(':id') @Roles('superuser','admin','instructor','provider','learner') one(@Param('id') id:string){const q=this.db.quizzes.find((x:any)=>x.id===id); if(!q) throw new NotFoundException('Quiz not found.'); return {success:true,data:q};}
 @Post() @Roles('superuser','admin','instructor','provider') @ApiBody({type:CreateQuizDto}) create(@Body() b:CreateQuizDto){ if(b.courseId && !this.db.ensureCourseExists(b.courseId)) throw new NotFoundException('Course not found.'); if(Array.isArray(b.questions)){ for(const question of b.questions){ if(!question.q || !Array.isArray(question.options) || question.options.length < 2) throw new BadRequestException('Question and options are required.'); if(question.ans === undefined || question.ans >= question.options.length) throw new BadRequestException('Correct answer must match one option.'); } } const q={id:this.db.id('q'),questions:[],...b,createdAt:new Date().toISOString()}; this.db.quizzes.push(q); return {success:true,data:q};}
 @Patch(':id') @Roles('superuser','admin','instructor','provider') @ApiBody({type:UpdateQuizDto}) update(@Param('id') id:string,@Body() b:UpdateQuizDto){const q=this.db.quizzes.find((x:any)=>x.id===id); if(!q) throw new NotFoundException('Quiz not found.'); if(b.courseId && !this.db.ensureCourseExists(b.courseId)) throw new NotFoundException('Course not found.'); if(Array.isArray(b.questions)){ for(const question of b.questions){ if(!question.q || !Array.isArray(question.options) || question.options.length < 2) throw new BadRequestException('Question and options are required.'); if(question.ans === undefined || question.ans >= question.options.length) throw new BadRequestException('Correct answer must match one option.'); } } Object.assign(q,b); return {success:true,data:q};}
 @Delete(':id') @Roles('superuser','admin','instructor','provider') remove(@Param('id') id:string){const i=this.db.quizzes.findIndex((x:any)=>x.id===id); if(i===-1) throw new NotFoundException('Quiz not found.'); this.db.quizzes.splice(i,1); return {success:true,data:{id}};}
 @Post(':id/submit') @Roles('superuser','learner') @ApiOperation({summary:'Submit quiz answers'}) @ApiBody({type:SubmitQuizDto}) submit(@Param('id') id:string,@Body() b:SubmitQuizDto){const q=this.db.quizzes.find((x:any)=>x.id===id); if(!q) throw new NotFoundException('Quiz not found.'); if(!this.db.ensureLearnerExists(b.learnerId)) throw new NotFoundException('Learner not found.'); const answers=b.answers||[]; const score=(q.questions||[]).reduce((s:any,question:any,i:number)=>s+(answers[i]===question.ans?1:0),0); return {success:true,data:{quizId:id,learnerId:b.learnerId,score,total:(q.questions||[]).length}};}
}